refactor(notes): component header note edit
