+++ src/modules/notes/NoteEdit/components/HeaderNoteEdit/FormNoteEdit.tsx
@@ -1,0 +1,8 @@
+import React, { memo } from "react";
+import FormNoteEditView from "./FormNoteEditView";
+
+const FormNoteEdit: React.FC = () => {
+	return <FormNoteEditView />;
+};
+
+export default memo(FormNoteEdit);
