--- src/modules/notes/NoteEdit/components/HeaderNoteEdit/HeaderNoteEdit.tsx
@@ -1,8 +1,0 @@
-import React, { memo } from "react";
-import HeaderNoteEditView from "./HeaderNoteEditView";
-
-const HeaderNoteEdit: React.FC = () => {
-	return <HeaderNoteEditView />;
-};
-
-export default memo(HeaderNoteEdit);
