+++ src/modules/notes/NoteEdit/components/HeaderNoteEdit/index.tsx
@@ -1,1 +1,1 @@
-export { default } from "./HeaderNoteEdit";
+export { default } from "./FormNoteEdit";
